requirements.txt
black
flake8
PyYAML
mypy
pytest